add stories to footer
