+++ src/components/PageTemplate/PageTemplate.tsx
@@ -2,7 +2,7 @@ import { PhaseBanner } from "@/components/govuk/PhaseBanner";
 import { Header } from "@/components/Header";
 import { Suspense } from "react";
 import CouncilConfigComponent from "@/config/CouncilConfigComponent";
-import { Footer } from "@/components/govuk/Footer";
+import { Footer } from "@/components/Footer";
 import { AppConfig } from "@/config/types";
 
 export interface PageTemplateProps {
